requirements.txt
djangorestframework
django==3.1.4


